summarization/requirements.txt
langchain-community==0.0.25
langchain-core==0.1.29
langchain-openai==0.0.8
langsmith==0.1.12
pypdf==4.1.0
